+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.11.3] - 2022-06-06
+
+### Fixed
+* **SharpDPAPI** project
+    * `masterkeys` command now accepts a `/password:X` argument with `/target:X`
+
+
 ## [1.11.2] - 2022-01-12
 
 ### Fixed